+++ react-app/src/hooks/index.js
@@ -9,6 +9,7 @@ import { SEARCH } from '@/constants';
 export { default as useRandomItems } from './useRandomItems';
 export { default as useRecentItems } from './useRecentItems';
 export { default as useRecentManager } from './useRecentManager';
+export { default as useServiceWorker } from './useServiceWorker';
 
 // useLocalStorage hook
 export const useLocalStorage = (key, initialValue) => {